# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Go untuk manga library: scan/move folder, bookmark, translate, bug report, duplikat.\nLogin lewat POST /login, lalu klik Authorize dan isi `Bearer \u003caccess_token\u003e`.",
        "title": "Manga Library API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/bookmarks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "List bookmark",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Item per halaman",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "object",
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/Bookmark.BookmarkRes"
                                                    }
                                                },
                                                "limit": {
                                                    "type": "integer"
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "pages": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hapus kalau sudah di-bookmark (200), buat kalau belum (201).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Toggle bookmark manga",
                "parameters": [
                    {
                        "description": "new_folders.id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "folder_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "folder_id": {
                                    "type": "integer"
                                },
                                "is_bookmarked": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "created_at": {
                                    "type": "string"
                                },
                                "folder_id": {
                                    "type": "integer"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "is_bookmarked": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bookmarks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Detail bookmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bookmarks.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Bookmark.Bookmark"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Hapus bookmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bookmarks.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bug-reports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bug-reports"
                ],
                "summary": "List bug report",
                "parameters": [
                    {
                        "enum": [
                            "all",
                            "open",
                            "fixed"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Filter status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "newest",
                            "oldest"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "Urutan",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.BugReportItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bug-reports"
                ],
                "summary": "Laporkan bug pada manga",
                "parameters": [
                    {
                        "description": "Laporan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBugReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/bug-reports/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bug-reports"
                ],
                "summary": "Ubah status bug report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "bug_reports.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status: open|fixed",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateBugReportStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Bookmark.Bookmark": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "folder": {
                    "$ref": "#/definitions/NewFolder.NewFolder"
                },
                "folder_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "Bookmark.BookmarkRes": {
            "type": "object",
            "properties": {
                "bookmark_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "integer"
                },
                "folder_name": {
                    "type": "string"
                },
                "folder_thumbnail": {
                    "type": "string"
                }
            }
        },
...
        "Model.BaseResponseModel": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {},
                "HeaderMessage": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
        "NewFolder.NewFolder": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_completed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
...
        "dto.BugReportItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "integer"
                },
                "folder_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateBugReportRequest": {
            "type": "object",
            "required": [
                "description",
                "folder_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateBugReportStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "description": "\"open\" | \"fixed\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: `Bearer \u003caccess_token\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}